chore(content-group): update base class to new layout
As we finished migrating all content patterns to the new layout aligning
to non-gutter grid, updates the base classes to the new one, that
finishes the migration.

Refs #4174.
