+++ packages/web-components/src/components/callout-with-media/callout-with-media.ts
@@ -1,7 +1,7 @@
 /**
  * @license
  *
- * Copyright IBM Corp. 2019, 2020
+ * Copyright IBM Corp. 2019, 2021
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -11,7 +11,7 @@ import { customElement } from 'lit-element';
 import ddsSettings from '@carbon/ibmdotcom-utilities/es/utilities/settings/settings.js';
 import styles from './callout-with-media.scss';
 import DDSCalloutMixin from '../../component-mixins/callout/callout';
-import DDSContentBlockSimple from '../content-block-simple/content-block-simple';
+import DDSContentBlock from '../content-block/content-block';
 
 const { stablePrefix: ddsPrefix } = ddsSettings;
 
@@ -21,7 +21,7 @@ const { stablePrefix: ddsPrefix } = ddsSettings;
  * @element dds-callout-data
  */
 @customElement(`${ddsPrefix}-callout-with-media`)
-class DDSCalloutWithMedia extends DDSCalloutMixin(DDSContentBlockSimple) {
+class DDSCalloutWithMedia extends DDSCalloutMixin(DDSContentBlock) {
   static styles = styles;
 }
