+++ packages/web-components/src/components/content-block-media/content-block-media.ts
@@ -9,7 +9,7 @@
 
 import { css, customElement } from 'lit-element';
 import ddsSettings from '@carbon/ibmdotcom-utilities/es/utilities/settings/settings.js';
-import DDSContentBlockSimple from '../content-block-simple/content-block-simple';
+import DDSContentBlock from '../content-block/content-block';
 import styles from './content-block-media.scss';
 
 const { stablePrefix: ddsPrefix } = ddsSettings;
@@ -20,7 +20,7 @@ const { stablePrefix: ddsPrefix } = ddsSettings;
  * @element dds-content-block-media
  */
 @customElement(`${ddsPrefix}-content-block-media`)
-class DDSContentBlockMedia extends DDSContentBlockSimple {
+class DDSContentBlockMedia extends DDSContentBlock {
   static get stableSelector() {
     return `${ddsPrefix}--content-block-media`;
   }
