+++ packages/web-components/src/components/content-block-mixed/content-block-mixed.ts
@@ -9,7 +9,7 @@
 
 import { customElement, css } from 'lit-element';
 import ddsSettings from '@carbon/ibmdotcom-utilities/es/utilities/settings/settings.js';
-import DDSContentBlockSimple from '../content-block-simple/content-block-simple';
+import DDSContentBlock from '../content-block/content-block';
 import styles from './content-block-mixed.scss';
 
 const { stablePrefix: ddsPrefix } = ddsSettings;
@@ -20,7 +20,7 @@ const { stablePrefix: ddsPrefix } = ddsSettings;
  * @element dds-content-block-mixed
  */
 @customElement(`${ddsPrefix}-content-block-mixed`)
-class DDSContentBlockMixed extends DDSContentBlockSimple {
+class DDSContentBlockMixed extends DDSContentBlock {
   static get stableSelector() {
     return `${ddsPrefix}--content-block-mixed`;
   }
