+++ packages/web-components/src/components/content-block-simple/content-block-simple.ts
@@ -7,16 +7,12 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import { Part } from 'lit-html';
-import { classMap } from 'lit-html/directives/class-map';
-import { html, css, customElement, TemplateResult } from 'lit-element';
-import settings from 'carbon-components/es/globals/js/settings.js';
+import { css, customElement } from 'lit-element';
 import ddsSettings from '@carbon/ibmdotcom-utilities/es/utilities/settings/settings.js';
 import StableSelectorMixin from '../../globals/mixins/stable-selector';
-import DDSContentBlock, { CONTENT_BLOCK_COMPLEMENTARY_STYLE_SCHEME } from '../content-block/content-block';
+import DDSContentBlock from '../content-block/content-block';
 import styles from './content-block-simple.scss';
 
-const { prefix } = settings;
 const { stablePrefix: ddsPrefix } = ddsSettings;
 
 /**
@@ -26,72 +22,6 @@ const { stablePrefix: ddsPrefix } = ddsSettings;
  */
 @customElement(`${ddsPrefix}-content-block-simple`)
 class DDSContentBlockSimple extends StableSelectorMixin(DDSContentBlock) {
-  /**
-   * The CSS class list for the container (grid) node.
-   */
-  protected _getContainerClasses(): string | ((part: Part) => void) {
-    const { complementaryStyleScheme, _hasComplementary: hasComplementary } = this;
-    return classMap({
-      [`${prefix}--content-layout`]: true,
-      [`${prefix}--content-layout--with-complementary`]: hasComplementary,
-      [`${prefix}--layout--border`]:
-        hasComplementary && complementaryStyleScheme === CONTENT_BLOCK_COMPLEMENTARY_STYLE_SCHEME.WITH_BORDER,
-    });
-  }
-
-  /**
-   * @returns The non-header, non-complementary contents.
-   */
-  protected _renderBody(): TemplateResult | string | void {
-    const { _hasContent: hasContent, _hasCopy: hasCopy, _hasMedia: hasMedia } = this;
-    return html`
-      <div ?hidden="${!hasContent && !hasCopy && !hasMedia}" class="${prefix}--content-layout__body">
-        ${super._renderBody()}
-      </div>
-    `;
-  }
-
-  /**
-   * @returns The media content.
-   */
-  protected _renderMedia(): TemplateResult | string | void {
-    const { _handleSlotChange: handleSlotChange } = this;
-    return html`
-      <slot name="media" @slotchange="${handleSlotChange}"></slot>
-    `;
-  }
-
-  protected _renderInnerBody() {
-    return html`
-      ${this._renderContent()}${this._renderMedia()}
-    `;
-  }
-
-  protected _renderContent(): TemplateResult | string | void {
-    const { _handleSlotChange: handleSlotChange } = this;
-    return html`
-      <slot @slotchange="${handleSlotChange}"></slot>
-    `;
-  }
-
-  protected _renderFooter(): TemplateResult | string | void {
-    const { _hasFooter: hasFooter, _handleSlotChange: handleSlotChange } = this;
-    // TODO: See if we can remove the surrounding `<div>`
-    return html`
-      <div ?hidden="${!hasFooter}">
-        <slot name="footer" @slotchange="${handleSlotChange}"></slot>
-      </div>
-    `;
-  }
-
-  render() {
-    return html`
-      <div class="${this._getContainerClasses()}">
-        ${this._renderHeading()}${this._renderBody()}${this._renderComplementary()}
-      </div>
-    `;
-  }
-
   static get stableSelector() {
     return `${ddsPrefix}--content-block-simple`;
   }
