+++ packages/web-components/src/components/content-group-simple/content-group-simple.ts
@@ -7,15 +7,12 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import { Part } from 'lit-html';
-import { html, css, customElement, TemplateResult } from 'lit-element';
-import settings from 'carbon-components/es/globals/js/settings.js';
+import { css, customElement } from 'lit-element';
 import ddsSettings from '@carbon/ibmdotcom-utilities/es/utilities/settings/settings.js';
 import StableSelectorMixin from '../../globals/mixins/stable-selector';
 import DDSContentGroup from '../content-group/content-group';
 import styles from './content-group-simple.scss';
 
-const { prefix } = settings;
 const { stablePrefix: ddsPrefix } = ddsSettings;
 
 /**
@@ -25,50 +22,6 @@ const { stablePrefix: ddsPrefix } = ddsSettings;
  */
 @customElement(`${ddsPrefix}-content-group-simple`)
 class DDSContentGroupSimple extends StableSelectorMixin(DDSContentGroup) {
-  /**
-   * The CSS class list for the container (grid) node.
-   */
-  // eslint-disable-next-line class-methods-use-this
-  protected _getContainerClasses(): string | ((part: Part) => void) {
-    return `${prefix}--content-layout`;
-  }
-
-  /**
-   * @returns The non-header, non-complementary contents.
-   */
-  protected _renderBody(): TemplateResult | string | void {
-    const { _hasContent: hasContent, _hasCopy: hasCopy, _hasMedia: hasMedia } = this;
-    return html`
-      <div ?hidden="${!hasContent && !hasCopy && !hasMedia}" class="${prefix}--content-layout__body">
-        ${super._renderBody()}
-      </div>
-    `;
-  }
-
-  protected _renderInnerBody(): TemplateResult | string | void {
-    return html`
-      ${this._renderMedia()}${this._renderContent()}
-    `;
-  }
-
-  protected _renderFooter(): TemplateResult | string | void {
-    const { _hasFooter: hasFooter, _handleSlotChange: handleSlotChange } = this;
-    // TODO: See if we can remove the surrounding `<div>`
-    return html`
-      <div ?hidden="${!hasFooter}">
-        <slot name="footer" @slotchange="${handleSlotChange}"></slot>
-      </div>
-    `;
-  }
-
-  render() {
-    return html`
-      <div class="${this._getContainerClasses()}">
-        ${this._renderHeading()}${this._renderBody()}${this._renderComplementary()}
-      </div>
-    `;
-  }
-
   static get stableSelector() {
     return `${ddsPrefix}--content-group-simple`;
   }
