+++ packages/web-components/src/components/content-group/content-group.ts
@@ -7,12 +7,8 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import { html, TemplateResult } from 'lit-element';
-import settings from 'carbon-components/es/globals/js/settings.js';
 import DDSContentBlock from '../content-block/content-block';
 import styles from './content-group.scss';
-
-const { prefix } = settings;
 
 /**
  * Content group.
@@ -20,26 +16,6 @@ const { prefix } = settings;
  * @abstract
  */
 class DDSContentGroup extends DDSContentBlock {
-  protected _renderFooter(): TemplateResult | string | void {
-    const { _hasFooter: hasFooter, _handleSlotChange: handleSlotChange } = this;
-    return html`
-      <div ?hidden="${!hasFooter}" class="${prefix}--content-group__cta-row">
-        <div class="${prefix}--content-group__cta ${prefix}--content-group__cta-col">
-          <slot name="footer" @slotchange="${handleSlotChange}"></slot>
-        </div>
-      </div>
-    `;
-  }
-
-  protected _renderInnerBody(): TemplateResult | string | void {
-    const { _hasContent: hasContent, _hasMedia: hasMedia, _handleSlotChange: handleSlotChange } = this;
-    return html`
-      <div ?hidden="${!hasContent && !hasMedia}" class="${prefix}--content-group__children ${prefix}--content-group__col">
-        <slot @slotchange="${handleSlotChange}"></slot>
-      </div>
-    `;
-  }
-
   static styles = styles; // `styles` here is a `CSSResult` generated by custom WebPack loader
 }
